rework types package
